Chore: #27 - MovieListVC 파일 분리
